working accelerometer and light/sensor tweaks
fixes a bug where the accelerometer did not enable due to passing a reference
to a short as the IOCTL parameter, when the IOCTL was attempting to dereference
it as an unsigned long, causing the value to change based on stale memory, resulting
in an invalid response.  this fixes that so the accelerometer will enable/disable
properly.

also added proper ALS value modification for the lights driver, based on information
found within the sholes kernel source for the lm3530 header and code.

finally tweaked the convert values for the orientation and accelerometer to make them
appear more sane (for instance orientation now doesn't register 460 degrees, and
accelerometer will appropriately register 9.8m/s^2 when at rest for gravity).  these
values should likely be tweaked more, but they seem to work now.

also reduced the default delay time on the accelerometer to 100ms, to speed up
orientation detection

Change-Id: I57f6031025e5ed4abff1a82dd745cd12169b0e00