Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

This messes up the accelerometer.  When orienting my phone horizontally the screen is reoriented but upside-down.  Also, i'm not certain what the LSG value is supposed to represent but when at rest the accelerometer should read on one axis an acceleration of ~9.8m/s^2 (to account for gravity), and with this commit it reports 13.6.

I will do more look into verifying correctness of the values and orientation presented, but for now this commit breaks reorientation of the phone.

Patch-set: 2