Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Though the maximum observed value is 5316, I've set the range to 27000 as that is the range reported by the official 2.2 sensor lib. 1400 was certainly too low. Even the array in the config.xml overlay specifies an autobrightness value for interval > 1500.

Patch-set: 1