Update patch set 1
Patch Set 1:

I agree that the free space should be increased but not through locales. This must be done through things that you can always download if necessary, such as the TTC, ringtones, wallpapers. Additional locales can only be added by making a custom build, but not all users can do it.

Patch-set: 1