Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The idea of using a base is to build on top of it, so I don't think the solution is to go to full_base and remove items from it.

If anything we should simply re-add the locales to sholes on top of small_base, or have a medium_base as Steve has suggested.  Personally I feel that medium_base is unnecessary, as I think distinguishing "standard" and "low space" devices is enough, and declaring a "medium space" device class might be cumbersome.

It was the intention of going to small_base that if we found anything missing that was desired, we would re-add them manually to the sholes config, as is the standard practice, so I would say go that route.

Patch-set: 1