Update patch set 1
Patch Set 1:

Personally I don't think that if we create a "medium_base" that it should be created exclusively with sholes in mind.  It would need to be generic so it can be used for other "medium space" devices.

However, as I have mentioned previously, I don't really see a need for a separate medium_base at this time.

Patch-set: 1