Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

As Chris has stated, the main problem is that you based this on your previous patch, which reverts back to full_base.

The proper solution should be to maintain the include of small_base, and then set the locales as you have done.

Patch-set: 1