Update patch set 1
Patch Set 1:

This patch doesn't actually modify the back vs. front facing microphone aspect of the audio policy, but instead removes some other parts of it.  It seems to be working fine on my copy of sholes, but I want to test some more before making a decision on this.

Patch-set: 1