Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I do however wonder if this value shouldn't be a CM standard as it is one of the few properties that is present in all devices (I couldn't find an exception, sorry if I missed one/them). 

I would defer this to Steve Kondik ... What do you think?

Patch-set: 1