Update patch set 1
Patch Set 1:

I think it's probably OK, but I'm not sure how this is used and if it's only scanning when the screen is on and not in LPM (pretty sure this is the case). If it's only when the screen is on, I think the power savings would be pretty tiny.

Patch-set: 1