Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

i'm going to go ahead and approve this as changing to 45 shouldn't be too bad of a change.  i have my doubts about the increase in battery life, but the other devices i compared against had their values set to either 45 or 90, so i don't think this change would be inappropriate.

Patch-set: 1