Update patch set 2
Patch Set 2:

Sorry for the incorrect 'torch-flash' suggestion on the forum; it appears the correct value is 'spotlight' as you have here now.

I confirm that the supplied patch a9e88dd55b289147078f5f97412cae0c28da7746 resolves the issue. (I have compiled, loaded, and tested it locally.) Torch is functional, both through the app and through the notification power widget.

(A previous comment of mine suggested that there was a link between this issue and a keyboard backlight issue.  That was a mistaken statement.  The Torch issue *IS* resolved by this patch, though!)

Patch-set: 2